refactor: extract UaSession from UaClient
Pure structural split with no behavior change:
- new asyncua/client/ua_session.py: owns authentication_token, subscription
  callbacks, publish loop, and AbstractSession service set
- UaClient retains transport (UASocketProtocol), secure channel, and the
  sessionless services (Discovery, Register/Unregister Server)
- UaClient holds a default UaSession in self._session and delegates the
  AbstractSession methods to it, so existing callers keep working

Also lands a few targeted fixes pulled from review of the prior squash:
- close_secure_channel fails in-flight RPC futures with ConnectionError
  before clearing _callbackmap, instead of leaving silent CancelledError
- delete_subscriptions builds the id set once with consistent int keys
- removed redundant protocol.check_answer in publish (send_request already
  checks); minor docstring/comment cleanup
