+++ asyncua/client/ua_client.py
@@ -1,12 +1,17 @@
 """
-Low level binary client
+Low level binary client.
+
+`UaClient` provides:
+- transport: socket and secure channel via `UASocketProtocol`,
+- sessionless services (Discovery, Register/Unregister Server),
+- a default `UaSession` accessible via `_session`, and back-compat delegating methods
+  for code that used to call session services directly on `UaClient`.
 """
 
 from __future__ import annotations
 
 import asyncio
 import copy
-import inspect
 import logging
 from collections.abc import Awaitable, Callable
 from typing import Any
@@ -20,8 +25,8 @@ from ..common.connection import SecureConnection, TransportLimits
 from ..common.utils import wait_for
 from ..crypto import security_policies
 from ..ua.ua_binary import header_from_binary, nodeid_from_binary, struct_from_binary, struct_to_binary, uatcp_to_binary
-from ..ua.uaerrors import BadNoSubscription, BadSessionClosed, BadTimeout, BadUserAccessDenied, UaStructParsingError
-from ..ua.uaprotocol_auto import OpenSecureChannelResult, SubscriptionAcknowledgement
+from ..ua.uaprotocol_auto import OpenSecureChannelResult
+from .ua_session import UaSession
 
 
 class UASocketProtocol(asyncio.Protocol):
@@ -291,21 +296,22 @@ class UASocketProtocol(asyncio.Protocol):
         self.logger.info("close_secure_channel")
         request = ua.CloseSecureChannelRequest()
         future = self._send_request(request, message_type=ua.MessageType.SecureClose)
-        # don't expect any more answers
         future.cancel()
+        # Fail any in-flight RPC futures with ConnectionError so awaiters can clean up,
+        # rather than leaving them as silently cancelled.
+        for pending in self._callbackmap.values():
+            if not pending.done():
+                pending.set_exception(ConnectionError("Secure channel closed"))
         self._callbackmap.clear()
-        # some servers send a response here, most do not ... so we ignore
 
 
 class UaClient(AbstractSession):
     """
-    low level OPC-UA client.
+    Low level OPC-UA client.
 
-    It implements (almost) all methods defined in asyncua spec
-    taking in argument the structures defined in asyncua spec.
-
-    In this Python implementation  most of the structures are defined in
-    uaprotocol_auto.py and uaprotocol_hand.py available under asyncua.ua
+    Owns the transport (`UASocketProtocol`) and a default `UaSession`.
+    Sessionless services (Discovery, Register/Unregister Server) live here.
+    Session-scoped services delegate to `self._session` for back-compat.
     """
 
     def __init__(self, timeout: float = 1.0) -> None:
@@ -313,13 +319,13 @@ class UaClient(AbstractSession):
         :param timeout: Timout in seconds
         """
         self.logger = logging.getLogger(f"{__name__}.UaClient")
-        self._subscription_callbacks: dict[int, Callable[..., Any]] = {}
         self._timeout = timeout
         self.security_policy: security_policies.SecurityPolicy = security_policies.SecurityPolicyNone()
         self.protocol: UASocketProtocol | None = None
-        self._publish_task: asyncio.Task[None] | None = None
         self._pre_request_hook: Callable[[], Awaitable[None]] | None = None
-        self._closing: bool = False
+        self._session: UaSession = UaSession(self)
+
+    # --- transport helpers ---
 
     def set_security(self, policy: security_policies.SecurityPolicy) -> None:
         self.security_policy = policy
@@ -342,7 +348,7 @@ class UaClient(AbstractSession):
     async def connect_socket(self, host: str, port: int) -> None:
         """Connect to server socket."""
         self.logger.info("opening connection")
-        self._closing = False
+        self._session._closing = False
         # Timeout the connection when the server isn't available
         await asyncio.wait_for(
             asyncio.get_running_loop().create_connection(self._make_protocol, host, port), self._timeout
@@ -351,7 +357,7 @@ class UaClient(AbstractSession):
     def disconnect_socket(self) -> None:
         if not self.protocol:
             return
-        if self.protocol and self.protocol.state == UASocketProtocol.CLOSED:
+        if self.protocol.state == UASocketProtocol.CLOSED:
             self.logger.warning("disconnect_socket was called but connection is closed")
             return
         self.protocol.disconnect_socket()
@@ -368,79 +374,10 @@ class UaClient(AbstractSession):
         return await self.protocol.open_secure_channel(params)
 
     async def close_secure_channel(self) -> None:
-        """
-        close secure channel. It seems to trigger a shutdown of socket
-        in most servers, so be prepared to reconnect
-        """
         if not self.protocol or self.protocol.state == UASocketProtocol.CLOSED:
             self.logger.warning("close_secure_channel was called but connection is closed")
             return None
         return await self.protocol.close_secure_channel()
-
-    async def create_session(self, parameters: ua.CreateSessionParameters) -> ua.CreateSessionResult:
-        if self.protocol is None:
-            raise ConnectionError("Connection is not open")
-        self.logger.info("create_session")
-        self._closing = False
-        # FIXME: setting a value on an object to set it its state is suspicious,
-        # especially when that object has its own state
-        self.protocol.closed = False
-        request = ua.CreateSessionRequest()
-        request.Parameters = parameters
-        data = await self._send_request(request)
-        response = struct_from_binary(ua.CreateSessionResponse, data)
-        self.logger.debug(response)
-        response.ResponseHeader.ServiceResult.check()
-        self.protocol.authentication_token = response.Parameters.AuthenticationToken
-        return response.Parameters
-
-    async def activate_session(self, parameters: ua.ActivateSessionParameters) -> ua.ActivateSessionResult:
-        self.logger.info("activate_session")
-        request = ua.ActivateSessionRequest()
-        request.Parameters = parameters
-        data = await self._send_request(request)
-        response = struct_from_binary(ua.ActivateSessionResponse, data)
-        self.logger.debug(response)
-        response.ResponseHeader.ServiceResult.check()
-        return response.Parameters
-
-    async def close_session(self, delete_subscriptions: bool) -> None:
-        self.logger.info("close_session")
-        if not self.protocol:
-            self.logger.warning("close_session but connection wasn't established")
-            return
-        self.protocol.closed = True
-        self._closing = True
-        if self._publish_task and not self._publish_task.done():
-            self._publish_task.cancel()
-        if self.protocol and self.protocol.state == UASocketProtocol.CLOSED:
-            self.logger.warning("close_session was called but connection is closed")
-            return
-        request = ua.CloseSessionRequest()
-        request.DeleteSubscriptions = delete_subscriptions
-        data = await self._send_request(request)
-        response = struct_from_binary(ua.CloseSessionResponse, data)
-        try:
-            response.ResponseHeader.ServiceResult.check()
-        except BadSessionClosed:
-            # Problem: closing the session with open publish requests leads to BadSessionClosed responses
-            #          we can just ignore it therefore.
-            #          Alternatively we could make sure that there are no publish requests in flight when
-            #          closing the session.
-            pass
-        except BadUserAccessDenied:
-            # Problem: older versions of asyncua didn't allow closing non-activated sessions. just ignore it.
-            pass
-
-    async def browse(self, parameters: ua.BrowseParameters) -> list[ua.BrowseResult]:
-        self.logger.info("browse")
-        request = ua.BrowseRequest()
-        request.Parameters = parameters
-        data = await self._send_request(request)
-        response = struct_from_binary(ua.BrowseResponse, data)
-        self.logger.debug(response)
-        response.ResponseHeader.ServiceResult.check()
-        return response.Results
 
     async def _send_request(
         self, request: Any, timeout: float = 1, message_type: ua.MessageType = ua.MessageType.SecureMessage
@@ -449,35 +386,36 @@ class UaClient(AbstractSession):
             raise ConnectionError("Connection is not open")
         return await self.protocol.send_request(request, timeout, message_type)
 
-    async def browse_next(self, parameters: ua.BrowseNextParameters) -> list[ua.BrowseResult]:
-        self.logger.debug("browse next")
-        request = ua.BrowseNextRequest()
-        request.Parameters = parameters
-        data = await self._send_request(request)
-        response = struct_from_binary(ua.BrowseNextResponse, data)
-        self.logger.debug(response)
-        response.ResponseHeader.ServiceResult.check()
-        return response.Parameters.Results
+    # --- back-compat: properties that previously lived on UaClient ---
 
-    async def read(self, parameters: ua.ReadParameters) -> list[ua.DataValue]:
-        self.logger.debug("read")
-        request = ua.ReadRequest()
-        request.Parameters = parameters
-        data = await self._send_request(request)
-        response = struct_from_binary(ua.ReadResponse, data)
-        self.logger.debug(response)
-        response.ResponseHeader.ServiceResult.check()
-        return response.Results
+    @property
+    def _subscription_callbacks(self) -> dict[int, Callable[..., Any]]:
+        return self._session._subscription_callbacks
 
-    async def write(self, params: ua.WriteParameters) -> list[ua.StatusCode]:
-        self.logger.debug("write")
-        request = ua.WriteRequest()
-        request.Parameters = params
-        data = await self._send_request(request)
-        response = struct_from_binary(ua.WriteResponse, data)
-        self.logger.debug(response)
-        response.ResponseHeader.ServiceResult.check()
-        return response.Results
+    @property
+    def _publish_task(self) -> asyncio.Task[None] | None:
+        return self._session._publish_task
+
+    @property
+    def _closing(self) -> bool:
+        return self._session._closing
+
+    @_closing.setter
+    def _closing(self, value: bool) -> None:
+        self._session._closing = value
+
+    # --- session lifecycle: delegate to default session ---
+
+    async def create_session(self, parameters: ua.CreateSessionParameters) -> ua.CreateSessionResult:
+        return await self._session.create_session(parameters)
+
+    async def activate_session(self, parameters: ua.ActivateSessionParameters) -> ua.ActivateSessionResult:
+        return await self._session.activate_session(parameters)
+
+    async def close_session(self, delete_subscriptions: bool) -> None:
+        return await self._session.close_session(delete_subscriptions)
+
+    # --- sessionless services ---
 
     async def get_endpoints(self, params: ua.GetEndpointsParameters) -> list[ua.EndpointDescription]:
         self.logger.debug("get_endpoint")
@@ -517,7 +455,6 @@ class UaClient(AbstractSession):
         response = struct_from_binary(ua.RegisterServerResponse, data)
         self.logger.debug(response)
         response.ResponseHeader.ServiceResult.check()
-        # nothing to return for this service
 
     async def unregister_server(self, registered_server: ua.RegisteredServer) -> None:
         self.logger.debug("unregister_server")
@@ -527,7 +464,6 @@ class UaClient(AbstractSession):
         response = struct_from_binary(ua.RegisterServerResponse, data)
         self.logger.debug(response)
         response.ResponseHeader.ServiceResult.check()
-        # nothing to return for this service
 
     async def register_server2(self, params: ua.RegisterServer2Parameters) -> list[ua.StatusCode]:
         self.logger.debug("register_server2")
@@ -549,278 +485,34 @@ class UaClient(AbstractSession):
         response.ResponseHeader.ServiceResult.check()
         return response.ConfigurationResults
 
+    # --- AbstractSession: delegate to default session ---
+
+    async def browse(self, parameters: ua.BrowseParameters) -> list[ua.BrowseResult]:
+        return await self._session.browse(parameters)
+
+    async def browse_next(self, parameters: ua.BrowseNextParameters) -> list[ua.BrowseResult]:
+        return await self._session.browse_next(parameters)
+
     async def translate_browsepaths_to_nodeids(self, browse_paths: list[ua.BrowsePath]) -> list[ua.BrowsePathResult]:
-        self.logger.debug("translate_browsepath_to_nodeid")
-        request = ua.TranslateBrowsePathsToNodeIdsRequest()
-        request.Parameters.BrowsePaths = browse_paths
-        data = await self._send_request(request)
-        response = struct_from_binary(ua.TranslateBrowsePathsToNodeIdsResponse, data)
-        self.logger.debug(response)
-        response.ResponseHeader.ServiceResult.check()
-        return response.Results
-
-    async def create_subscription(  # type: ignore[override]
-        self, params: ua.CreateSubscriptionParameters, callback: Callable[..., Any]
-    ) -> ua.CreateSubscriptionResult:
-        self.logger.debug("create_subscription")
-        request = ua.CreateSubscriptionRequest()
-        request.Parameters = params
-        data = await self._send_request(request)
-        response = struct_from_binary(ua.CreateSubscriptionResponse, data)
-        response.ResponseHeader.ServiceResult.check()
-        self._subscription_callbacks[response.Parameters.SubscriptionId] = callback
-        self.logger.info("create_subscription success SubscriptionId %s", response.Parameters.SubscriptionId)
-        if not self._publish_task or self._publish_task.done():
-            # Start the publishing loop if it is not yet running
-            # The current strategy is to have only one open publish request per UaClient. This might not be enough
-            # in high latency networks or in case many subscriptions are created. A Set of Tasks of `_publish_loop`
-            # could be used if necessary.
-            self._publish_task = asyncio.create_task(self._publish_loop())
-        return response.Parameters
-
-    async def inform_subscriptions(self, status: ua.StatusCode) -> None:
-        """
-        Inform all current subscriptions with a status code. This calls the handler's status_change_notification
-        """
-        status_message = ua.StatusChangeNotification(Status=status)
-        notification_message = ua.NotificationMessage(NotificationData=[status_message])  # type: ignore[list-item]
-        for subid, callback in self._subscription_callbacks.items():
-            try:
-                parameters = ua.PublishResult(subid, NotificationMessage=notification_message)
-                if inspect.iscoroutinefunction(callback):
-                    await callback(parameters)
-                else:
-                    callback(parameters)
-            except Exception:  # we call user code, catch everything!
-                self.logger.exception("Exception while calling user callback: %s")
-
-    async def update_subscription(self, params: ua.ModifySubscriptionParameters) -> ua.ModifySubscriptionResult:
-        request = ua.ModifySubscriptionRequest()
-        request.Parameters = params
-        data = await self._send_request(request)
-        response = struct_from_binary(ua.ModifySubscriptionResponse, data)
-        response.ResponseHeader.ServiceResult.check()
-        self.logger.info("update_subscription success SubscriptionId %s", params.SubscriptionId)
-        return response.Parameters
-
-    modify_subscription = update_subscription  # legacy support
-
-    async def delete_subscriptions(self, subscription_ids: list[int]) -> list[ua.StatusCode]:  # type: ignore[override]
-        self.logger.debug("delete_subscriptions %r", subscription_ids)
-        request = ua.DeleteSubscriptionsRequest()
-        request.Parameters.SubscriptionIds = subscription_ids
-        data = await self._send_request(request)
-        response = struct_from_binary(ua.DeleteSubscriptionsResponse, data)
-        response.ResponseHeader.ServiceResult.check()
-        self.logger.info("remove subscription callbacks for %r", subscription_ids)
-        for sid in subscription_ids:
-            self._subscription_callbacks.pop(sid)
-        return response.Results
-
-    async def publish(self, acks: list[ua.SubscriptionAcknowledgement]) -> ua.PublishResponse:
-        """
-        Send a PublishRequest to the server.
-        """
-        self.logger.debug("publish %r", acks)
-        request = ua.PublishRequest()
-        request.Parameters.SubscriptionAcknowledgements = acks if acks else []
-        data = await self._send_request(request, timeout=0)
-        assert self.protocol is not None
-        self.protocol.check_answer(data, "while waiting for publish response")
-        try:
-            response = struct_from_binary(ua.PublishResponse, data)
-        except Exception as ex:
-            self.logger.exception("Error parsing notification from server")
-            raise UaStructParsingError from ex
-        return response
-
-    async def _publish_loop(self) -> None:
-        """
-        Start a loop that sends a publish requests and waits for the publish responses.
-        Forward the `PublishResult` to the matching `Subscription` by callback.
-        """
-        ack: SubscriptionAcknowledgement | None = None
-        while not self._closing:
-            try:
-                response = await self.publish([ack] if ack else [])
-            except BadTimeout:  # See Spec. Part 4, 7.28
-                # Repeat without acknowledgement
-                ack = None
-                continue
-            except BadNoSubscription:  # See Spec. Part 5, 13.8.1
-                # BadNoSubscription is expected to be received after deleting the last subscription.
-                # We use this as a signal to exit this task and stop sending PublishRequests. This is easier than
-                # checking if there are no more subscriptions registered in this client (). A Publish response
-                # could still arrive before the DeleteSubscription response.
-                #
-                # We could remove the callback already when sending the DeleteSubscription request,
-                # but there are some legitimate reasons to keep them around, such as when the server
-                # responds with "BadTimeout" and we should try again later instead of just removing
-                # the subscription client-side.
-                #
-                # There are a variety of ways to act correctly, but the most practical solution seems
-                # to be to just silently ignore any BadNoSubscription responses.
-                self.logger.info("BadNoSubscription received, ignoring because it's probably valid.")
-                # End task
-                return
-            except UaStructParsingError:
-                ack = None
-                continue
-            subscription_id = response.Parameters.SubscriptionId
-            if not subscription_id:
-                # The value 0 is used to indicate that there were no Subscriptions defined for which a
-                # response could be sent. See Spec. Part 4 - Section 5.13.5 "Publish"
-                # End task
-                return
-            try:
-                callback = self._subscription_callbacks[subscription_id]
-            except KeyError:
-                self.logger.warning(
-                    "Received data for unknown subscription %s active are %s",
-                    subscription_id,
-                    self._subscription_callbacks.keys(),
-                )
-            else:
-                try:
-                    if inspect.iscoroutinefunction(callback):
-                        await callback(response.Parameters)
-                    else:
-                        callback(response.Parameters)
-                except Exception:  # we call user code, catch everything!
-                    self.logger.exception("Exception while calling user callback: %s")
-            # Repeat with acknowledgement
-            if response.Parameters.NotificationMessage.NotificationData:
-                ack = ua.SubscriptionAcknowledgement()
-                ack.SubscriptionId = subscription_id
-                ack.SequenceNumber = response.Parameters.NotificationMessage.SequenceNumber
-            else:
-                ack = None
-
-    async def create_monitored_items(
-        self, params: ua.CreateMonitoredItemsParameters
-    ) -> list[ua.MonitoredItemCreateResult]:
-        self.logger.info("create_monitored_items")
-        request = ua.CreateMonitoredItemsRequest()
-        request.Parameters = params
-        data = await self._send_request(request)
-        response = struct_from_binary(ua.CreateMonitoredItemsResponse, data)
-        self.logger.debug(response)
-        response.ResponseHeader.ServiceResult.check()
-        return response.Results
-
-    async def delete_monitored_items(self, params: ua.DeleteMonitoredItemsParameters) -> list[ua.StatusCode]:
-        self.logger.info("delete_monitored_items")
-        request = ua.DeleteMonitoredItemsRequest()
-        request.Parameters = params
-        data = await self._send_request(request)
-        response = struct_from_binary(ua.DeleteMonitoredItemsResponse, data)
-        self.logger.debug(response)
-        response.ResponseHeader.ServiceResult.check()
-        return response.Results
-
-    async def add_nodes(self, nodestoadd: list[ua.AddNodesItem]) -> list[ua.AddNodesResult]:
-        self.logger.info("add_nodes")
-        request = ua.AddNodesRequest()
-        request.Parameters.NodesToAdd = nodestoadd
-        data = await self._send_request(request)
-        response = struct_from_binary(ua.AddNodesResponse, data)
-        self.logger.debug(response)
-        response.ResponseHeader.ServiceResult.check()
-        return response.Results
-
-    async def add_references(self, refs: list[ua.AddReferencesItem]) -> list[ua.StatusCode]:
-        self.logger.info("add_references")
-        request = ua.AddReferencesRequest()
-        request.Parameters.ReferencesToAdd = refs
-        data = await self._send_request(request)
-        response = struct_from_binary(ua.AddReferencesResponse, data)
-        self.logger.debug(response)
-        response.ResponseHeader.ServiceResult.check()
-        return response.Results
-
-    async def delete_references(self, refs: list[ua.DeleteReferencesItem]) -> list[ua.StatusCode]:
-        self.logger.info("delete")
-        request = ua.DeleteReferencesRequest()
-        request.Parameters.ReferencesToDelete = refs
-        data = await self._send_request(request)
-        response = struct_from_binary(ua.DeleteReferencesResponse, data)
-        self.logger.debug(response)
-        response.ResponseHeader.ServiceResult.check()
-        return response.Parameters.Results
-
-    async def delete_nodes(self, params: ua.DeleteNodesParameters) -> list[ua.StatusCode]:
-        self.logger.info("delete_nodes")
-        request = ua.DeleteNodesRequest()
-        request.Parameters = params
-        data = await self._send_request(request)
-        response = struct_from_binary(ua.DeleteNodesResponse, data)
-        self.logger.debug(response)
-        response.ResponseHeader.ServiceResult.check()
-        return response.Results
-
-    async def call(self, methodstocall: list[ua.CallMethodRequest]) -> list[ua.CallMethodResult]:
-        request = ua.CallRequest()
-        request.Parameters.MethodsToCall = methodstocall
-        data = await self._send_request(request)
-        response = struct_from_binary(ua.CallResponse, data)
-        self.logger.debug(response)
-        response.ResponseHeader.ServiceResult.check()
-        return response.Results
-
-    async def history_read(self, params: ua.HistoryReadParameters) -> list[ua.HistoryReadResult]:
-        self.logger.info("history_read")
-        request = ua.HistoryReadRequest()
-        request.Parameters = params
-        data = await self._send_request(request)
-        response = struct_from_binary(ua.HistoryReadResponse, data)
-        self.logger.debug(response)
-        response.ResponseHeader.ServiceResult.check()
-        return response.Results
-
-    async def modify_monitored_items(
-        self, params: ua.ModifyMonitoredItemsParameters
-    ) -> list[ua.MonitoredItemModifyResult]:
-        self.logger.info("modify_monitored_items")
-        request = ua.ModifyMonitoredItemsRequest()
-        request.Parameters = params
-        data = await self._send_request(request)
-        response = struct_from_binary(ua.ModifyMonitoredItemsResponse, data)
-        self.logger.debug(response)
-        response.ResponseHeader.ServiceResult.check()
-        return response.Results
+        return await self._session.translate_browsepaths_to_nodeids(browse_paths)
 
     async def register_nodes(self, nodes: list[ua.NodeId]) -> list[ua.NodeId]:
-        self.logger.info("register_nodes")
-        request = ua.RegisterNodesRequest()
-        request.Parameters.NodesToRegister = nodes
-        data = await self._send_request(request)
-        response = struct_from_binary(ua.RegisterNodesResponse, data)
-        self.logger.debug(response)
-        response.ResponseHeader.ServiceResult.check()
-        return response.Parameters.RegisteredNodeIds
+        return await self._session.register_nodes(nodes)
 
     async def unregister_nodes(self, nodes: list[ua.NodeId]) -> None:
-        self.logger.info("unregister_nodes")
-        request = ua.UnregisterNodesRequest()
-        request.Parameters.NodesToUnregister = nodes
-        data = await self._send_request(request)
-        response = struct_from_binary(ua.UnregisterNodesResponse, data)
-        self.logger.debug(response)
-        response.ResponseHeader.ServiceResult.check()
-        # nothing to return for this service
+        return await self._session.unregister_nodes(nodes)
+
+    async def read(self, parameters: ua.ReadParameters) -> list[ua.DataValue]:
+        return await self._session.read(parameters)
+
+    async def write(self, params: ua.WriteParameters) -> list[ua.StatusCode]:
+        return await self._session.write(params)
+
+    async def history_read(self, params: ua.HistoryReadParameters) -> list[ua.HistoryReadResult]:
+        return await self._session.history_read(params)
 
     async def read_attributes(self, nodeids: list[ua.NodeId], attr: ua.AttributeIds) -> list[ua.DataValue]:
-        self.logger.info("read_attributes of several nodes")
-        request = ua.ReadRequest()
-        for nodeid in nodeids:
-            rv = ua.ReadValueId()
-            rv.NodeId = nodeid
-            rv.AttributeId = attr
-            request.Parameters.NodesToRead.append(rv)
-        data = await self._send_request(request)
-        response = struct_from_binary(ua.ReadResponse, data)
-        response.ResponseHeader.ServiceResult.check()
-        return response.Results
+        return await self._session.read_attributes(nodeids, attr)
 
     async def write_attributes(
         self,
@@ -828,50 +520,60 @@ class UaClient(AbstractSession):
         datavalues: list[ua.DataValue],
         attributeid: ua.AttributeIds = ua.AttributeIds.Value,
     ) -> list[ua.StatusCode]:
-        """
-        Set an attribute of multiple nodes
-        datavalue is a ua.DataValue object
-        """
-        self.logger.info("write_attributes of several nodes")
-        request = ua.WriteRequest()
-        for idx, nodeid in enumerate(nodeids):
-            attr = ua.WriteValue()
-            attr.NodeId = nodeid
-            attr.AttributeId = attributeid
-            attr.Value = datavalues[idx]
-            request.Parameters.NodesToWrite.append(attr)
-        data = await self._send_request(request)
-        response = struct_from_binary(ua.WriteResponse, data)
-        response.ResponseHeader.ServiceResult.check()
-        return response.Results
+        return await self._session.write_attributes(nodeids, datavalues, attributeid)
 
-    async def set_monitoring_mode(self, params: ua.SetMonitoringModeParameters) -> list[ua.uatypes.StatusCode]:
-        """
-        Update the subscription monitoring mode
-        """
-        self.logger.info("set_monitoring_mode")
-        request = ua.SetMonitoringModeRequest()
-        request.Parameters = params
-        data = await self._send_request(request)
-        response = struct_from_binary(ua.SetMonitoringModeResponse, data)
-        self.logger.debug(response)
-        response.ResponseHeader.ServiceResult.check()
-        return response.Parameters.Results
+    async def add_nodes(self, nodestoadd: list[ua.AddNodesItem]) -> list[ua.AddNodesResult]:
+        return await self._session.add_nodes(nodestoadd)
 
-    async def set_publishing_mode(self, params: ua.SetPublishingModeParameters) -> list[ua.uatypes.StatusCode]:
-        """
-        Update the subscription publishing mode
-        """
-        self.logger.info("set_publishing_mode")
-        request = ua.SetPublishingModeRequest()
-        request.Parameters = params
-        data = await self._send_request(request)
-        response = struct_from_binary(ua.SetPublishingModeResponse, data)
-        self.logger.debug(response)
-        response.ResponseHeader.ServiceResult.check()
-        return response.Parameters.Results
+    async def add_references(self, refs: list[ua.AddReferencesItem]) -> list[ua.StatusCode]:
+        return await self._session.add_references(refs)
+
+    async def delete_references(self, refs: list[ua.DeleteReferencesItem]) -> list[ua.StatusCode]:
+        return await self._session.delete_references(refs)
+
+    async def delete_nodes(self, params: ua.DeleteNodesParameters) -> list[ua.StatusCode]:
+        return await self._session.delete_nodes(params)
+
+    async def call(self, methodstocall: list[ua.CallMethodRequest]) -> list[ua.CallMethodResult]:
+        return await self._session.call(methodstocall)
+
+    async def create_subscription(  # type: ignore[override]
+        self, params: ua.CreateSubscriptionParameters, callback: Callable[..., Any]
+    ) -> ua.CreateSubscriptionResult:
+        return await self._session.create_subscription(params, callback)
+
+    async def update_subscription(self, params: ua.ModifySubscriptionParameters) -> ua.ModifySubscriptionResult:
+        return await self._session.update_subscription(params)
+
+    modify_subscription = update_subscription  # legacy support
+
+    async def delete_subscriptions(self, subscription_ids: list[int]) -> list[ua.StatusCode]:  # type: ignore[override]
+        return await self._session.delete_subscriptions(subscription_ids)
 
     async def transfer_subscriptions(self, params: ua.TransferSubscriptionsParameters) -> list[ua.TransferResult]:
-        # Subscriptions aren't bound to a Session and can be transferred!
-        # https://reference.opcfoundation.org/Core/Part4/v104/5.13.7/
-        raise NotImplementedError
+        return await self._session.transfer_subscriptions(params)
+
+    async def inform_subscriptions(self, status: ua.StatusCode) -> None:
+        return await self._session.inform_subscriptions(status)
+
+    async def publish(self, acks: list[ua.SubscriptionAcknowledgement]) -> ua.PublishResponse:
+        return await self._session.publish(acks)
+
+    async def create_monitored_items(
+        self, params: ua.CreateMonitoredItemsParameters
+    ) -> list[ua.MonitoredItemCreateResult]:
+        return await self._session.create_monitored_items(params)
+
+    async def delete_monitored_items(self, params: ua.DeleteMonitoredItemsParameters) -> list[ua.StatusCode]:
+        return await self._session.delete_monitored_items(params)
+
+    async def modify_monitored_items(
+        self, params: ua.ModifyMonitoredItemsParameters
+    ) -> list[ua.MonitoredItemModifyResult]:
+        return await self._session.modify_monitored_items(params)
+
+    async def set_monitoring_mode(self, params: ua.SetMonitoringModeParameters) -> list[ua.uatypes.StatusCode]:
+        return await self._session.set_monitoring_mode(params)
+
+    async def set_publishing_mode(self, params: ua.SetPublishingModeParameters) -> list[ua.uatypes.StatusCode]:
+        return await self._session.set_publishing_mode(params)
